LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-import-parent-pseudo.tentative-expected.txt

Inside
Inner (W)
Inner (U)

FAIL The & selector matches the scoping root assert_equals: expected "1" but got ""
FAIL The & selector behaves like :scope assert_equals: expected "1" but got ""
